Made a screensaver out of this.
Changed a little bit the code of the MainForm in order to make of
BinaryClock a screensaver for Windows. :-)
It will display a preview in the screensaver selection screen of
Windows, and it will do its job when you PC is down (you get what I
mean). To quit it, don't move the mouse, press a key. ;-)